# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Compsole API documentation.\n\n### Authenticating\n\nThere are two methods of authenticating to the Compsole API. **Basic Auth** is used solely for the purpose of the Compsole UI. **Api Key Auth** is used for service accounts to authenticate prior to accessing the REST endpoints.\n\n#### Basic Auth\n\nUse the [`/auth/local/login`](#operations-Auth_API-post_auth_local_login) endpoint below to authenticate from the Compsole UI.\n\n#### Api Key Auth\n\nAPI Key Authentication is more complicated.\n\n1. You must retreive your `api_key` and `api_secret` from the Compsole UI after creating a service account.\n2. Use the [`/rest/token`](#operations-Auth_API-post_rest_login) endpoint to retrieve an API Token to use in requests\n3. Place the API Token into the `Authorization` header like so: `Authorization: Bearer \u003capi token here...\u003e`\n\n##### Refresh Tokens\n\nRefresh tokens can be used to renew a session without re-authenticating. The refresh token should be set in the `refresh-token` cookie already, so you can simply make a `POST` request to `/rest/token/refresh` and receive a new API token from this endpoint.\n",
        "title": "Compsole API",
        "contact": {
            "name": "BradHacker",
            "url": "http://github.com/BradHacker/compsole/issues"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
...
    },
    "basePath": "/api",
    "paths": {
...
        "/rest/provider": {
            "get": {
                "security": [
                    {
                        "ServiceAuth": []
                    }
                ],
                "description": "List all Providers",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Service API"
                ],
                "summary": "List all Providers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/rest.ProviderModel"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ServiceAuth": []
                    }
                ],
                "description": "Create a Provider",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Service API"
                ],
                "summary": "Create a Provider",
                "parameters": [
                    {
                        "description": "The provider to create",
                        "name": "provider",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest.ProviderInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/rest.ProviderModel"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    }
                }
            }
        },
        "/rest/provider/{id}": {
            "get": {
                "security": [
                    {
                        "ServiceAuth": []
                    }
                ],
                "description": "Get a provider",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Service API"
                ],
                "summary": "Get a provider",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "example": "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx",
                        "description": "The id of the provider",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.ProviderModel"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ServiceAuth": []
                    }
                ],
                "description": "Update a Provider",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Service API"
                ],
                "summary": "Update a Provider",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "example": "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx",
                        "description": "The id of the provider",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "The updated provider",
                        "name": "provider",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest.ProviderInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/rest.ProviderModel"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ServiceAuth": []
                    }
                ],
                "description": "Delete a Provider",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Service API"
                ],
                "summary": "Delete a Provider",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "example": "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx",
                        "description": "The id of the provider",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.APIError": {
            "type": "object",
            "properties": {
                "error": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "rest.CompetitionEdge": {
            "description": "Used for Competition in edges",
            "type": "object",
            "properties": {
                "id": {
                    "description": "Fields",
                    "type": "string",
                    "example": "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
                },
                "name": {
                    "description": "[REQUIRED] The unique name (aka. slug) for the competition.",
                    "type": "string",
                    "example": "Test Competition"
                }
            }
        },
...
        "rest.ProviderInput": {
            "description": "Used as an input model for creating/updating Providers",
            "type": "object",
            "required": [
                "config",
                "name",
                "type"
            ],
            "properties": {
                "config": {
                    "description": "See https://github.com/BradHacker/compsole/tree/main/configs for examples",
                    "type": "string",
                    "example": "See https://github.com/BradHacker/compsole/tree/main/configs for examples"
                },
                "name": {
                    "type": "string",
                    "example": "RITSEC Openstack"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "OPENSTACK"
                    ],
                    "example": "OPENSTACK"
                }
            }
        },
        "rest.ProviderModel": {
            "description": "Used for Provider endpoints",
            "type": "object",
            "properties": {
                "config": {
                    "description": "[REQUIRED] This is the JSON configuration for the provider.",
                    "type": "string",
                    "example": "{...}"
                },
                "id": {
                    "description": "Fields",
                    "type": "string",
                    "example": "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
                },
                "name": {
                    "description": "[REQUIRED] The unique name (aka. slug) for the provider.",
                    "type": "string",
                    "example": "Local Openstack"
                },
                "provider_to_competitions": {
                    "description": "Edges",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rest.CompetitionEdge"
                    }
                },
                "type": {
                    "description": "[REQUIRED] The type of provider this is (must match a registered one in https://github.com/BradHacker/compsole/tree/main/compsole/providers)",
                    "type": "string",
                    "example": "OPENSTACK"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ServiceAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "UserAuth": {
            "type": "basic"
        }
    },
...
}